rename gateway API package
On-behalf-of: @SAP christoph.mewes@sap.com
